moved common types into common.ts
